+++ paddle/fluid/operators/collective/c_comm_init_all_op.cu.cc
@@ -1,4 +1,4 @@
-/* Copyright (c) 2019 PaddlePaddle Authors. All Rights Reserved.
+/* Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,34 +14,7 @@ limitations under the License. */
 
 #include "paddle/fluid/operators/collective/c_comm_init_all_op.h"
 
-#if defined(PADDLE_WITH_NCCL) || defined(PADDLE_WITH_RCCL)
-#include "paddle/fluid/platform/device/gpu/nccl_helper.h"
-#endif
-
-namespace paddle {
-namespace operators {
-
-template <typename T, typename DeviceContext>
-class CCommInitAllKernel : public framework::OpKernel<T> {
- public:
-  void Compute(const framework::ExecutionContext& ctx) const override {
-#if defined(PADDLE_WITH_NCCL) || defined(PADDLE_WITH_RCCL)
-    std::vector<int> devices = ctx.Attr<std::vector<int>>("devices");
-    if (devices.empty()) {
-      devices = platform::GetSelectedDevices();
-    }
-
-    int rid = ctx.Attr<int>("ring_id");
-
-    platform::NCCLCommContext::Instance().CreateAllNCCLComms(devices, rid);
-#endif
-  }
-};
-
-}  // namespace operators
-}  // namespace paddle
-
 namespace ops = paddle::operators;
 
 PD_REGISTER_STRUCT_KERNEL(
-    c_comm_init_all, GPU, ALL_LAYOUT, ops::CCommInitAllKernel, float, double) {}
+    c_comm_init_all, GPU, ALL_LAYOUT, ops::CCommInitAllKernel, float) {}
